--- scheduling/shared/scheduling_exports.lua
@@ -1,12 +1,0 @@
-
-local scheduling = {}
-
-
-scheduling.delay = require("shared.delay")
-
-scheduling.nextTick = require("shared.next_tick")
-
-scheduling.skip = require("shared.skip")
-
-
-umg.expose("scheduling", scheduling)
