--- sync/shared/sync_exports.lua
@@ -1,36 +1,0 @@
-
-
-local sync = {}
-
-
-sync.proxyEventToClient = require("shared.proxy")
-
-sync.autoSyncComponent = require("shared.auto_sync_component")
-
-
-local tickDelta = require("shared.tick_delta")
-sync.getTickDelta = tickDelta.getTickDelta
-sync.getTimeOfLastTick = tickDelta.getTimeOfLastTick
-
-
-local control = require("shared.control")
-
-sync.isControlledBy = control.isControlledBy
-
-if client then
-    -- only available clientside
-    sync.isClientControlling = control.isClientControlling
-end
-
-
-
-if server then
-    -- only available serverside.
-    sync.syncComponent = require("shared.manual_sync_component")
-end
-
-
-
-
-
-umg.expose("sync", sync)
