+++ ui/client/exports.lua
@@ -1,19 +1,16 @@
 
-
-
-
-
-
 local ui = {}
 
 ui.elements = require("client.elements")
----@type fun(x?:number,y?:number,w?:number,h?:number):ui.Region
 ui.Region = require("kirigami.Region")
----@type fun(name:string):(ElementClass|Element)
 ui.Element = require("client.newElement")
 
 
+if false then
+    ---Provides user interface functionality.
+    ---
+    ---Availability: **Client**
+    _G.ui = ui
+end
 umg.expose("ui", ui)
-
-
-
+return ui
